updating DUMB
From the repository here: https://github.com/kode54/dumb

Someone made fixes for ARM platforms and other nice things.

+++ dumb/include/internal/riff.h
@@ -3,22 +3,20 @@
 
 struct riff;
 
-struct riff_chunk
-{
-	unsigned type;
-    int32 offset;
-	unsigned size;
-    struct riff * nested;
+struct riff_chunk {
+    unsigned type;
+    long offset;
+    unsigned size;
+    struct riff *nested;
 };
 
-struct riff
-{
-	unsigned type;
-	unsigned chunk_count;
-	struct riff_chunk * chunks;
+struct riff {
+    unsigned type;
+    unsigned chunk_count;
+    struct riff_chunk *chunks;
 };
 
-struct riff * riff_parse( DUMBFILE * f, int32 offset, int32 size, unsigned proper );
-void riff_free( struct riff * );
+struct riff *riff_parse(DUMBFILE *f, long offset, long size, unsigned proper);
+void riff_free(struct riff *);
 
 #endif
